Move NTP config command to separate file
