+++ config/main.py
@@ -46,6 +46,7 @@ from . import vxlan
 from . import plugins
 from .config_mgmt import ConfigMgmtDPB, ConfigMgmt
 from . import mclag
+from . import ntp
 
 # mock masic APIs for unit test
 try:
@@ -1143,6 +1144,7 @@ config.add_command(flow_counters.flowcnt_route)
 config.add_command(kdump.kdump)
 config.add_command(kube.kubernetes)
 config.add_command(muxcable.muxcable)
+config.add_command(ntp.ntp)
 
 if 'INCLUDE_NAT: n' not in exclude_cli_list:
     #add nat commands
@@ -5525,57 +5527,6 @@ def del_syslog_server(ctx, syslog_ip_address):
     except SystemExit as e:
         ctx.fail("Restart service rsyslog-config failed with error {}".format(e))
 
-#
-# 'ntp' group ('config ntp ...')
-#
-@config.group(cls=clicommon.AbbreviationGroup)
-@click.pass_context
-def ntp(ctx):
-    """NTP server configuration tasks"""
-    config_db = ConfigDBConnector()
-    config_db.connect()
-    ctx.obj = {'db': config_db}
-
-@ntp.command('add')
-@click.argument('ntp_ip_address', metavar='<ntp_ip_address>', required=True)
-@click.pass_context
-def add_ntp_server(ctx, ntp_ip_address):
-    """ Add NTP server IP """
-    if not clicommon.is_ipaddress(ntp_ip_address):
-        ctx.fail('Invalid ip address')
-    db = ctx.obj['db']
-    ntp_servers = db.get_table("NTP_SERVER")
-    if ntp_ip_address in ntp_servers:
-        click.echo("NTP server {} is already configured".format(ntp_ip_address))
-        return
-    else:
-        db.set_entry('NTP_SERVER', ntp_ip_address, {'NULL': 'NULL'})
-        click.echo("NTP server {} added to configuration".format(ntp_ip_address))
-        try:
-            click.echo("Restarting ntp-config service...")
-            clicommon.run_command("systemctl restart ntp-config", display_cmd=False)
-        except SystemExit as e:
-            ctx.fail("Restart service ntp-config failed with error {}".format(e))
-
-@ntp.command('del')
-@click.argument('ntp_ip_address', metavar='<ntp_ip_address>', required=True)
-@click.pass_context
-def del_ntp_server(ctx, ntp_ip_address):
-    """ Delete NTP server IP """
-    if not clicommon.is_ipaddress(ntp_ip_address):
-        ctx.fail('Invalid IP address')
-    db = ctx.obj['db']
-    ntp_servers = db.get_table("NTP_SERVER")
-    if ntp_ip_address in ntp_servers:
-        db.set_entry('NTP_SERVER', '{}'.format(ntp_ip_address), None)
-        click.echo("NTP server {} removed from configuration".format(ntp_ip_address))
-    else:
-        ctx.fail("NTP server {} is not configured.".format(ntp_ip_address))
-    try:
-        click.echo("Restarting ntp-config service...")
-        clicommon.run_command("systemctl restart ntp-config", display_cmd=False)
-    except SystemExit as e:
-        ctx.fail("Restart service ntp-config failed with error {}".format(e))
 
 #
 # 'sflow' group ('config sflow ...')
